+++ sample/src/commonMain/kotlin/com/kez/picker/sample/ui/screen/BackgroundStylePickerScreen.kt
@@ -105,12 +105,14 @@ internal fun BackgroundStylePickerScreen(
                 ) {
                     YearMonthPicker(
                         state = yearMonthState,
-                        yearPickerLabel = "연도",
-                        monthPickerLabel = "월",
-                        yearItemContentDescription = { "${it}년" },
-                        monthItemContentDescription = { getMonthContentDescription(it) },
-                        previousItemActionLabel = "이전 항목 선택",
-                        nextItemActionLabel = "다음 항목 선택",
+                        accessibility = PickerDefaults.yearMonthPickerAccessibility(
+                            yearPickerLabel = "연도",
+                            monthPickerLabel = "월",
+                            yearItemContentDescription = { "${it}년" },
+                            monthItemContentDescription = { getMonthContentDescription(it) },
+                            previousItemActionLabel = "이전 항목 선택",
+                            nextItemActionLabel = "다음 항목 선택"
+                        ),
                         style = PickerDefaults.style(
                             textStyles = PickerDefaults.textStyles(
                                 textStyle = TextStyle(
@@ -144,14 +146,16 @@ internal fun BackgroundStylePickerScreen(
                 ) {
                     TimePicker(
                         state = timeState,
-                        hourPickerLabel = "시간",
-                        minutePickerLabel = "분",
-                        periodPickerLabel = "오전/오후",
-                        hourItemContentDescription = { "${it}시" },
-                        minuteItemContentDescription = { "${it}분" },
-                        periodItemContentDescription = { getTimePeriodContentDescription(it) },
-                        previousItemActionLabel = "이전 항목 선택",
-                        nextItemActionLabel = "다음 항목 선택",
+                        accessibility = PickerDefaults.timePickerAccessibility(
+                            hourPickerLabel = "시간",
+                            minutePickerLabel = "분",
+                            periodPickerLabel = "오전/오후",
+                            hourItemContentDescription = { "${it}시" },
+                            minuteItemContentDescription = { "${it}분" },
+                            periodItemContentDescription = { getTimePeriodContentDescription(it) },
+                            previousItemActionLabel = "이전 항목 선택",
+                            nextItemActionLabel = "다음 항목 선택"
+                        ),
                         style = PickerDefaults.style(
                             textStyles = PickerDefaults.textStyles(
                                 textStyle = TextStyle(